+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.x.y] - Unreleased
+
+- Added experimental options for Icarus Verilog linting.
+  - `verilog.linting.iverilog.includePath` is to specify include directory.
+  - `verilog.linting.iverilog.verilogHDL.standard` is to specify standard rules for Verilog-HDL files.
+  - `verilog.linting.iverilog.systemVerilog.standard` is to specify standard rules for SystemVerilog files.
+  - \[Caution\] I've only tested on Ubuntu on WSL2 platform. If you find any problems about Icarus Verilog linting, please let me know.
+
 ## [1.8.1] - 2022-12-26
 
 - Fixed `Instantiate Module` not working issue. [#376](https://github.com/mshr-h/vscode-verilog-hdl-support/issues/376)